+++ include/ArduinoJson/JsonVariantBase.hpp
@@ -13,6 +13,7 @@ namespace ArduinoJson {
 
 // Forward declarations.
 class JsonArraySubscript;
+template <typename TKey>
 class JsonObjectSubscript;
 
 template <typename TImpl>
@@ -76,8 +77,10 @@ class JsonVariantBase : public Internals::JsonPrintable<TImpl> {
   // Returns the value associated with the specified key if the variant is
   // an object.
   // Return JsonVariant::invalid() if the variant is not an object.
-  FORCE_INLINE const JsonObjectSubscript operator[](const char *key) const;
-  FORCE_INLINE const JsonObjectSubscript operator[](const String &key) const;
+  FORCE_INLINE const JsonObjectSubscript<const char *> operator[](
+      const char *key) const;
+  FORCE_INLINE const JsonObjectSubscript<const String &> operator[](
+      const String &key) const;
 
   // Serialize the variant to a JsonWriter
   void writeTo(Internals::JsonWriter &writer) const;