+++ test/ArduinoString_Tests.cpp
@@ -7,103 +7,160 @@
 #include <gtest/gtest.h>
 #include <ArduinoJson.h>
 
-TEST(ArduinoStringTests, JsonBuffer_ParseArray) {
-  DynamicJsonBuffer jsonBuffer;
-  String json("[1,2]");
-  JsonArray &array = jsonBuffer.parseArray(json);
+class ArduinoStringTests : public ::testing::Test {
+ protected:
+  static void eraseString(String &str) {
+    char *p = const_cast<char *>(str.c_str());
+    while (*p) *p++ = '*';
+  }
+
+  DynamicJsonBuffer _jsonBuffer;
+};
+
+TEST_F(ArduinoStringTests, JsonBuffer_ParseArray) {
+  String json("[\"hello\"]");
+  JsonArray &array = _jsonBuffer.parseArray(json);
+  eraseString(json);
   ASSERT_TRUE(array.success());
+  ASSERT_STREQ("hello", array[0]);
 }
 
-TEST(ArduinoStringTests, JsonBuffer_ParseObject) {
-  DynamicJsonBuffer jsonBuffer;
-  String json("{\"a\":1,\"b\":2}");
-  JsonObject &object = jsonBuffer.parseObject(json);
+TEST_F(ArduinoStringTests, JsonBuffer_ParseObject) {
+  String json("{\"hello\":\"world\"}");
+  JsonObject &object = _jsonBuffer.parseObject(json);
+  eraseString(json);
   ASSERT_TRUE(object.success());
+  ASSERT_STREQ("world", object["hello"]);
 }
 
-TEST(ArduinoStringTests, JsonVariant) {
-  String input = "Hello world!";
-  JsonVariant variant(input);
-  ASSERT_TRUE(variant.is<String>());
-  String output = variant.as<String>();
-  ASSERT_EQ(input, output);
-}
-
-TEST(ArduinoStringTests, JsonObject_Subscript) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_Subscript) {
   char json[] = "{\"key\":\"value\"}";
-  JsonObject &object = jsonBuffer.parseObject(json);
+  JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_STREQ("value", object[String("key")]);
 }
 
-TEST(ArduinoStringTests, JsonObject_ConstSubscript) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_ConstSubscript) {
   char json[] = "{\"key\":\"value\"}";
-  const JsonObject &object = jsonBuffer.parseObject(json);
+  const JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_STREQ("value", object[String("key")]);
 }
 
-TEST(ArduinoStringTests, JsonObject_Set) {
-  DynamicJsonBuffer jsonBuffer;
-  JsonObject &object = jsonBuffer.createObject();
-  String key = "key";
-  object.set(key, "value");
-  ASSERT_STREQ("value", object["key"]);
+TEST_F(ArduinoStringTests, JsonObject_SetKey) {
+  JsonObject &object = _jsonBuffer.createObject();
+  String key("hello");
+  object.set(key, "world");
+  eraseString(key);
+  ASSERT_STREQ("world", object["hello"]);
 }
 
-TEST(ArduinoStringTests, JsonObject_Get) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_SetValue) {
+  JsonObject &object = _jsonBuffer.createObject();
+  String value("world");
+  object.set("hello", value);
+  eraseString(value);
+  ASSERT_STREQ("world", object["hello"]);
+}
+
+TEST_F(ArduinoStringTests, JsonObject_SetKeyValue) {
+  JsonObject &object = _jsonBuffer.createObject();
+  String key("hello");
+  String value("world");
+  object.set(key, value);
+  eraseString(key);
+  eraseString(value);
+  ASSERT_STREQ("world", object["hello"]);
+}
+
+TEST_F(ArduinoStringTests, JsonObject_Get) {
   char json[] = "{\"key\":\"value\"}";
-  const JsonObject &object = jsonBuffer.parseObject(json);
+  const JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_STREQ("value", object.get(String("key")));
 }
 
-TEST(ArduinoStringTests, JsonObject_GetT) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_GetT) {
   char json[] = "{\"key\":\"value\"}";
-  const JsonObject &object = jsonBuffer.parseObject(json);
+  const JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_STREQ("value", object.get<const char *>(String("key")));
 }
 
-TEST(ArduinoStringTests, JsonObject_IsT) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_IsT) {
   char json[] = "{\"key\":\"value\"}";
-  const JsonObject &object = jsonBuffer.parseObject(json);
+  const JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_TRUE(object.is<const char *>(String("key")));
 }
 
-TEST(ArduinoStringTests, JsonObject_CreateNestedObject) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_CreateNestedObject) {
   String key = "key";
   char json[64];
-  JsonObject &object = jsonBuffer.createObject();
+  JsonObject &object = _jsonBuffer.createObject();
   object.createNestedObject(key);
+  eraseString(key);
   object.printTo(json, sizeof(json));
   ASSERT_STREQ("{\"key\":{}}", json);
 }
 
-TEST(ArduinoStringTests, JsonObject_CreateNestedArray) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_CreateNestedArray) {
   String key = "key";
   char json[64];
-  JsonObject &object = jsonBuffer.createObject();
+  JsonObject &object = _jsonBuffer.createObject();
   object.createNestedArray(key);
+  eraseString(key);
   object.printTo(json, sizeof(json));
   ASSERT_STREQ("{\"key\":[]}", json);
 }
 
-TEST(ArduinoStringTests, JsonObject_ContainsKey) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_ContainsKey) {
   char json[] = "{\"key\":\"value\"}";
-  const JsonObject &object = jsonBuffer.parseObject(json);
+  const JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_TRUE(object.containsKey(String("key")));
 }
 
-TEST(ArduinoStringTests, JsonObject_Remove) {
-  DynamicJsonBuffer jsonBuffer;
+TEST_F(ArduinoStringTests, JsonObject_Remove) {
   char json[] = "{\"key\":\"value\"}";
-  JsonObject &object = jsonBuffer.parseObject(json);
+  JsonObject &object = _jsonBuffer.parseObject(json);
   ASSERT_EQ(1, object.size());
   object.remove(String("key"));
   ASSERT_EQ(0, object.size());
+}
+
+TEST_F(ArduinoStringTests, JsonObjectSubscript_SetKey) {
+  JsonObject &object = _jsonBuffer.createObject();
+  String key("hello");
+  object[key] = "world";
+  eraseString(key);
+  ASSERT_STREQ("world", object["hello"]);
+}
+
+TEST_F(ArduinoStringTests, JsonObjectSubscript_SetValue) {
+  JsonObject &object = _jsonBuffer.createObject();
+  String value("world");
+  object["hello"] = value;
+  eraseString(value);
+  ASSERT_STREQ("world", object["hello"]);
+}
+
+TEST_F(ArduinoStringTests, JsonArray_Add) {
+  JsonArray &array = _jsonBuffer.createArray();
+  String value("hello");
+  array.add(value);
+  eraseString(value);
+  ASSERT_STREQ("hello", array[0]);
+}
+
+TEST_F(ArduinoStringTests, JsonArray_Set) {
+  JsonArray &array = _jsonBuffer.createArray();
+  String value("world");
+  array.add("hello");
+  array.set(0, value);
+  eraseString(value);
+  ASSERT_STREQ("world", array[0]);
+}
+
+TEST_F(ArduinoStringTests, JsonArraySubscript) {
+  JsonArray &array = _jsonBuffer.createArray();
+  String value("world");
+  array.add("hello");
+  array[0] = value;
+  eraseString(value);
+  ASSERT_STREQ("world", array[0]);
 }