+++ test/JsonObject_PrintTo_Tests.cpp
@@ -11,131 +11,105 @@ using namespace ArduinoJson::Internals;
 
 class JsonObject_PrintTo_Tests : public testing::Test {
  public:
-  JsonObject_PrintTo_Tests() : object(json.createObject()) {}
+  JsonObject_PrintTo_Tests() : _object(_jsonBuffer.createObject()) {}
 
  protected:
   void outputMustBe(const char *expected) {
     char actual[256];
-    size_t actualLen = object.printTo(actual, sizeof(actual));
-    size_t measuredLen = object.measureLength();
+    size_t actualLen = _object.printTo(actual, sizeof(actual));
+    size_t measuredLen = _object.measureLength();
 
     EXPECT_STREQ(expected, actual);
     EXPECT_EQ(strlen(expected), actualLen);
     EXPECT_EQ(strlen(expected), measuredLen);
   }
 
-  StaticJsonBuffer<JSON_OBJECT_SIZE(2)> json;
-  JsonObject &object;
+  DynamicJsonBuffer _jsonBuffer;
+  JsonObject &_object;
 };
 
 TEST_F(JsonObject_PrintTo_Tests, EmptyObject) { outputMustBe("{}"); }
 
-TEST_F(JsonObject_PrintTo_Tests, OneString) {
-  object["key"] = "value";
-
-  outputMustBe("{\"key\":\"value\"}");
-}
-
 TEST_F(JsonObject_PrintTo_Tests, TwoStrings) {
-  object["key1"] = "value1";
-  object["key2"] = "value2";
+  _object["key1"] = "value1";
+  _object.set("key2", "value2");
 
   outputMustBe("{\"key1\":\"value1\",\"key2\":\"value2\"}");
 }
 
 TEST_F(JsonObject_PrintTo_Tests, RemoveFirst) {
-  object["key1"] = "value1";
-  object["key2"] = "value2";
-  object.remove("key1");
+  _object["key1"] = "value1";
+  _object["key2"] = "value2";
+  _object.remove("key1");
 
   outputMustBe("{\"key2\":\"value2\"}");
 }
 
 TEST_F(JsonObject_PrintTo_Tests, RemoveLast) {
-  object["key1"] = "value1";
-  object["key2"] = "value2";
-  object.remove("key2");
+  _object["key1"] = "value1";
+  _object["key2"] = "value2";
+  _object.remove("key2");
 
   outputMustBe("{\"key1\":\"value1\"}");
 }
 
 TEST_F(JsonObject_PrintTo_Tests, RemoveUnexistingKey) {
-  object["key1"] = "value1";
-  object["key2"] = "value2";
-  object.remove("key3");
+  _object["key1"] = "value1";
+  _object["key2"] = "value2";
+  _object.remove("key3");
 
   outputMustBe("{\"key1\":\"value1\",\"key2\":\"value2\"}");
 }
 
 TEST_F(JsonObject_PrintTo_Tests, ReplaceExistingKey) {
-  object["key"] = "value1";
-  object["key"] = "value2";
+  _object["key"] = "value1";
+  _object["key"] = "value2";
 
   outputMustBe("{\"key\":\"value2\"}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneStringOverCapacity) {
-  object["key1"] = "value1";
-  object["key2"] = "value2";
-  object["key3"] = "value3";
-
-  outputMustBe("{\"key1\":\"value1\",\"key2\":\"value2\"}");
+TEST_F(JsonObject_PrintTo_Tests, TwoIntegers) {
+  _object["a"] = 1;
+  _object.set("b", 2);
+  outputMustBe("{\"a\":1,\"b\":2}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneInteger) {
-  object["key"] = 1;
-  outputMustBe("{\"key\":1}");
+TEST_F(JsonObject_PrintTo_Tests, TwoDoublesFourDigits) {
+  _object["a"] = double_with_n_digits(3.14159265358979323846, 4);
+  _object.set("b", 2.71828182845904523536, 4);
+  outputMustBe("{\"a\":3.1416,\"b\":2.7183}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneDoubleFourDigits) {
-  object["key"] = double_with_n_digits(3.14159265358979323846, 4);
-  outputMustBe("{\"key\":3.1416}");
+TEST_F(JsonObject_PrintTo_Tests, TwoDoubleDefaultDigits) {
+  _object["a"] = 3.14159265358979323846;
+  _object.set("b", 2.71828182845904523536);
+  outputMustBe("{\"a\":3.14,\"b\":2.72}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneDoubleDefaultDigits) {
-  object["key"] = 3.14159265358979323846;
-  outputMustBe("{\"key\":3.14}");
+TEST_F(JsonObject_PrintTo_Tests, TwoNull) {
+  _object["a"] = static_cast<char *>(0);
+  _object.set("b", static_cast<char *>(0));
+  outputMustBe("{\"a\":null,\"b\":null}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneNull) {
-  object["key"] = static_cast<char *>(0);
-  outputMustBe("{\"key\":null}");
+TEST_F(JsonObject_PrintTo_Tests, TwoBooleans) {
+  _object["a"] = true;
+  _object.set("b", false);
+  outputMustBe("{\"a\":true,\"b\":false}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneTrue) {
-  object["key"] = true;
-  outputMustBe("{\"key\":true}");
+TEST_F(JsonObject_PrintTo_Tests, ThreeNestedArrays) {
+  _object.createNestedArray("a");
+  _object["b"] = _jsonBuffer.createArray();
+  _object.set("c", _jsonBuffer.createArray());
+
+  outputMustBe("{\"a\":[],\"b\":[],\"c\":[]}");
 }
 
-TEST_F(JsonObject_PrintTo_Tests, OneFalse) {
-  object["key"] = false;
-  outputMustBe("{\"key\":false}");
-}
+TEST_F(JsonObject_PrintTo_Tests, ThreeNestedObjects) {
+  _object.createNestedObject("a");
+  _object["b"] = _jsonBuffer.createObject();
+  _object.set("c", _jsonBuffer.createObject());
 
-TEST_F(JsonObject_PrintTo_Tests, OneEmptyNestedArrayViaProxy) {
-  JsonArray &nestedArray = json.createArray();
-
-  object["key"] = nestedArray;
-
-  outputMustBe("{\"key\":[]}");
-}
-
-TEST_F(JsonObject_PrintTo_Tests, OneEmptyNestedObjectViaProxy) {
-  JsonObject &nestedArray = json.createObject();
-
-  object["key"] = nestedArray;
-
-  outputMustBe("{\"key\":{}}");
-}
-
-TEST_F(JsonObject_PrintTo_Tests, OneEmptyNestedObject) {
-  object.createNestedObject("key");
-
-  outputMustBe("{\"key\":{}}");
-}
-
-TEST_F(JsonObject_PrintTo_Tests, OneEmptyNestedArray) {
-  object.createNestedArray("key");
-
-  outputMustBe("{\"key\":[]}");
+  outputMustBe("{\"a\":{},\"b\":{},\"c\":{}}");
 }